Feat: add eventNetwork system
