+++ games/rtype/include/System/System.hpp
@@ -8,9 +8,9 @@
 #ifndef RTYPE_SYSTEM_HPP
 #define RTYPE_SYSTEM_HPP
 
-
 #include "CollisionSystem.hpp"
 #include "DispawnableSystem.hpp"
+#include "EventNetworkSystem.hpp"
 #include "EventSystem.hpp"
 #include "MoveSystem.hpp"
 #include "SpriteSystem.hpp"
